% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annif_api_info.R
\name{get_annif_api_info}
\alias{get_annif_api_info}
\title{Get basic information about the Annif REST API}
\usage{
get_annif_api_info()
}
\description{
This function retrieves basic information such as the title and version
of the Annif API service. It connects to the Annif REST API endpoint
and returns metadata about the API service.
}
\examples{
# Get API information
api_info <- get_annif_api_info()
print(api_info)
}
