% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annif_projects.R
\name{get_annif_projects}
\alias{get_annif_projects}
\title{Get a list of available projects}
\usage{
get_annif_projects()
}
\value{
A data frame (tibble) containing information about available
}
\description{
This function retrieves information about all available projects
in the Annif service, including their configuration, training status,
and associated vocabulary information.
}
\examples{
\donttest{
# Get all available projects
projects <- get_annif_projects()
print(projects)
}

}
