% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annif_project_info.R
\name{get_annif_vocabulary_info}
\alias{get_annif_vocabulary_info}
\title{Get a list of available vocabularies}
\usage{
get_annif_vocabulary_info()
}
\value{
A data frame (tibble) containing information about available
}
\description{
This function retrieves information about all available vocabularies
in the Annif service, including their IDs, languages, sizes, and
loading status.
}
\examples{
\donttest{
# Get all available vocabularies
vocabs <- get_annif_vocabulary_info()
print(vocabs)
}

}
